backend-apis/deployment_scripts/requirements.txt
google-cloud-aiplatform
google-cloud-discoveryengine
google-cloud-storage
google-cloud-firestore
